refactor code base (#23)
Refactor code base to make the plugin easier to use as well as improve
the overall functionality and extendability:

* Update code base to mostly match WordPress coding standards (e.g.
`array()` instead of `[]`).

* Update output buffer timing to start earlier on the `setup_theme`
hook instead of the `template_redirect` hook (thank you @noknokcody).
This is needed to increase compatibility with other plugins.

* Update settings:

    * Update CDN URL to be CDN Hostname. This change makes it so the
rewrite is only focused on rewriting the site hostname with the CDN
hostname, regardless of the scheme in the URL.

    * Remove Included Directories and replace it with CDN Inclusions. Add
File Extensions setting to allow URLs to be included in the rewrite
based on the file extension. The default file extensions added are
`.avif`, `.css`, `.gif`, `.jpeg`, `.jpg`, `.json`, `.js`, `.mp3`,
`.mp4`, `.pdf`, `.png`, `.svg`, and `.webp`.

    * Update Exclusions to be CDN Exclusions. Add Strings setting to allow
URLs to be excluded in the rewrite based on any string in the URL. This
is the same behavior as before but documented differently as previously
it could have been more than just directories and/or extensions that
were excluded. The currently limitation for this is query strings, but
this will be improved on in the future.

    * Remove Relative Paths setting in favor of
`cdn_enabler_rewrite_relative_urls` filter hook.

    * Remove CDN HTTPS setting in favor of replacing the site hostname with
the CDN hostname while inheriting the protocol that was already in
place.

    * Add configuration validation. This currently validates the CDN
Hostname to ensure it returns a successful `200` status code (by making
a `HEAD` request to a test file). This will make it so URLs are not
replaced with value that could cause errors. Extending this to also
validate the KeyCDN API Key and KeyCDN Zone ID will also be added
(similar to how it is done when purging the CDN cache).

    * Improve validation.

* Update requirements checks to also check for PHP version.

* Update translated strings to be escaped and to not contain HTML.

* Update purge CDN cache handling.

* Update KeyCDN notice to not be displayed if `.kxcdn.com` domain is
used.

* Add new rewrite engine that focuses on replacing the site hostname
with the CDN hostname, regardless of the scheme in the URL, based on if
the file extension in the URL is included. The idea behind this is any
file contents can be sent through it to be rewritten, whether that is
an HTML page or even CSS, JavaScript, or JSON files. This improves the
handling (#19) and fixes limitations (#14).

* Add WP-CLI cache purging through the `cdn-enabler` command and
`purge` subcommand. Though it is an uncommon configuration, this does
have a limitation for multisite networks that use more than one Zone as
it will only purge the CDN cache for the main site. This will be
expanded upon in the future when the CDN cache purging is improved for
multisite networks.

* Add `cdn_enabler_user_can_purge_cache`,
`cdn_enabler_page_contents_before_rewrite`,
`cdn_enabler_bypass_rewrite`, `cdn_enabler_site_hostnames`, and
`cdn_enabler_rewrite_relative_urls` filter hooks.

* Fix requirement notices being shown to all users.

* Deprecate `user_can_clear_cache` filter hook in favor of
replacement.

* Remove the `the_content` filter hook added PR #11 because the updated
rewrite handling now covers this in addition to other data that is
obtained through the WordPress API.

Fixes #14, fixes #19, and closes #21