fix purge cache request handling
Fix purge cache request handling to redirect to the current URL path
with the query parameters `_cache`, `_action`, `_wpnonce` removed
instead of `wp_get_referer()`. Previously, if the `Referer` request
header was not sent, like if `Referrer-Policy: no-referrer` response
header was set, this would cause the redirect to be cancelled and then
script to be terminated. This lead to the response body being empty.