update configuration validation (#42)
Update configuration validation to include the `home_url()` as an HTTP
`Referer` in the test request.  This URL is what the frontend uses and
will be the base HTTP `Referer` for the majority of requests to the
CDN. While some HTTP referrers may have the CDN hostname in them and
not just the site hostname, I do not think it is worth making a second
request to also check this. Furthermore, having the CDN hostname as an
allowed HTTP referrer it is not required if a website does not have
assets delivered by the CDN that are then making requests to other
assets delivered by the CDN. Updating this validation will prevent
rewriting URLs to use the CDN hostname if a `403` status would be
returned when being requested by the pages that have been rewritten.